+++ core-ng/src/main/java/core/framework/impl/kafka/KafkaMessageListener.java
@@ -1,6 +1,7 @@
 package core.framework.impl.kafka;
 
 import core.framework.impl.json.JSONReader;
+import core.framework.impl.log.LogManager;
 import core.framework.kafka.BulkMessageHandler;
 import core.framework.kafka.MessageHandler;
 import core.framework.util.Exceptions;
@@ -8,9 +9,14 @@ import core.framework.util.Maps;
 import core.framework.util.Sets;
 import core.framework.util.StopWatch;
 import org.apache.kafka.clients.consumer.Consumer;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.apache.kafka.common.serialization.ByteArrayDeserializer;
+import org.apache.kafka.common.serialization.StringDeserializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.time.Duration;
 import java.util.Map;
 import java.util.Set;
 
@@ -18,18 +24,30 @@ import java.util.Set;
  * @author neo
  */
 public class KafkaMessageListener {
+    public final ConsumerMetrics consumerMetrics;
     final Map<String, MessageHandlerHolder<?>> handlerHolders = Maps.newHashMap();
     final Map<String, BulkMessageHandlerHolder<?>> bulkHandlerHolders = Maps.newHashMap();
-    final Kafka kafka;
+    final LogManager logManager;
+
     private final Logger logger = LoggerFactory.getLogger(KafkaMessageListener.class);
+    private final String uri;
     private final String name;
     private final Set<String> topics = Sets.newHashSet();
+
     public int poolSize = Runtime.getRuntime().availableProcessors() * 4;
+    public Duration maxProcessTime = Duration.ofMinutes(30);
+    public int maxPollRecords = 500;            // default kafka setting, refer to org.apache.kafka.clients.consumer.ConsumerConfig.MAX_POLL_RECORDS_CONFIG
+    public int maxPollBytes = 3 * 1024 * 1024;  // get 3M bytes message at max
+    public int minPollBytes = 1;                // default kafka setting
+    public Duration minPollMaxWaitTime = Duration.ofMillis(500);
+
     private KafkaMessageListenerThread[] listenerThreads;
 
-    KafkaMessageListener(Kafka kafka, String name) {
-        this.kafka = kafka;
+    public KafkaMessageListener(String uri, String name, LogManager logManager) {
+        this.uri = uri;
         this.name = name;
+        this.logManager = logManager;
+        this.consumerMetrics = new ConsumerMetrics(name);
     }
 
     public <T> void subscribe(String topic, Class<T> messageClass, MessageHandler<T> handler, BulkMessageHandler<T> bulkHandler) {
@@ -43,26 +61,54 @@ public class KafkaMessageListener {
 
     public void start() {
         listenerThreads = new KafkaMessageListenerThread[poolSize];
-        String group = kafka.logManager.appName;
+        String group = logManager.appName;
         for (int i = 0; i < poolSize; i++) {
             StopWatch watch = new StopWatch();
             String name = "kafka-listener-" + (this.name == null ? "" : this.name + "-") + i;
-            Consumer<String, byte[]> consumer = kafka.consumer(group, topics);
+            Consumer<String, byte[]> consumer = consumer(group, topics);
             KafkaMessageListenerThread thread = new KafkaMessageListenerThread(name, consumer, this);
             thread.start();
             listenerThreads[i] = thread;
             logger.info("create kafka listener thread, name={}, topics={}, elapsedTime={}", name, topics, watch.elapsedTime());
         }
-        logger.info("kafka listener started, uri={}, topics={}, name={}", kafka.uri, topics, name);
+        logger.info("kafka listener started, uri={}, topics={}, name={}", uri, topics, name);
     }
 
-    public void stop() {
-        logger.info("stop kafka listener, uri={}, topics={}, name={}", kafka.uri, topics, name);
+    public void shutdown() {
+        logger.info("shutting down kafka listener, uri={}, name={}", uri, name);
         if (listenerThreads != null) {
             for (KafkaMessageListenerThread thread : listenerThreads) {
                 thread.shutdown();
             }
         }
+    }
+
+    public void awaitTermination() {
+        if (listenerThreads != null) {
+            for (KafkaMessageListenerThread thread : listenerThreads) {
+                thread.awaitTermination();
+            }
+        }
+        logger.info("kafka listener stopped, uri={}, topics={}, name={}", uri, topics, name);
+    }
+
+    private Consumer<String, byte[]> consumer(String group, Set<String> topics) {
+        if (uri == null) throw new Error("uri must not be null");
+        Map<String, Object> config = Maps.newHashMap();
+        config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, uri);
+        config.put(ConsumerConfig.GROUP_ID_CONFIG, group);
+        config.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, false);
+        config.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "latest");
+        config.put(ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG, (int) maxProcessTime.toMillis());
+        config.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, (int) maxProcessTime.plusSeconds(5).toMillis());
+        config.put(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, maxPollRecords);
+        config.put(ConsumerConfig.FETCH_MAX_BYTES_CONFIG, maxPollBytes);
+        config.put(ConsumerConfig.FETCH_MIN_BYTES_CONFIG, minPollBytes);
+        config.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, (int) minPollMaxWaitTime.toMillis());
+        Consumer<String, byte[]> consumer = new KafkaConsumer<>(config, new StringDeserializer(), new ByteArrayDeserializer());
+        consumer.subscribe(topics);
+        consumerMetrics.add(consumer.metrics());
+        return consumer;
     }
 
     static class BulkMessageHandlerHolder<T> {