+++ core-ng/src/main/java/core/framework/impl/web/service/WebServiceClient.java
@@ -31,9 +31,8 @@ import java.util.Map;
  * @author neo
  */
 public class WebServiceClient {
-    private final Logger logger = LoggerFactory.getLogger(WebServiceClient.class);
     public static final String USER_AGENT = "APIClient";
-
+    private final Logger logger = LoggerFactory.getLogger(WebServiceClient.class);
     private final String serviceURL;
     private final HTTPClient httpClient;
     private final RequestBeanMapper mapper;