+++ core-ng/src/main/java/core/framework/module/CacheConfig.java
@@ -82,7 +82,7 @@ public class CacheConfig extends Config {
         RedisImpl redis = new RedisImpl("redis-cache");
         redis.host = host;
         redis.timeout(Duration.ofSeconds(1));   // for cache, use shorter timeout than default redis config
-        context.shutdownHook.add(redis::close);
+        context.shutdownHook.methods.add(redis::close);
         context.backgroundTask().scheduleWithFixedDelay(redis.pool::refresh, Duration.ofMinutes(5));
         context.stat.metrics.add(new PoolMetrics(redis.pool));
         configureCacheManager(new RedisCacheStore(redis));