+++ core-ng/src/main/java/core/framework/module/DBConfig.java
@@ -43,7 +43,7 @@ public class DBConfig extends Config {
 
     private DatabaseImpl createDatabase() {
         DatabaseImpl database = new DatabaseImpl("db" + (name == null ? "" : "-" + name));
-        context.shutdownHook.add(database::close);
+        context.shutdownHook.methods.add(database::close);
         context.backgroundTask().scheduleWithFixedDelay(database.pool::refresh, Duration.ofMinutes(10));
         context.stat.metrics.add(new PoolMetrics(database.pool));
         context.beanFactory.bind(Database.class, name, database);