+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.3.7 (6/21/2018)
 * api: make RemoteServiceException exposes https status
 * site: provide default ajax/json error response handling to cover common cases 
+* make http server / scheduler / kafka listener / executor shutdown gracefully 
 
 ### 6.3.6 (6/14/2018 - 6/21/2018)
 * api: assert same service interface must not have duplicate method name